# specs/001-flight-zone-map/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: DroneGo Flight Zone Map API
  description: |
    REST API for UK drone flight restriction zone data and TOAL (Take Off and Landing) site information.
    Provides geospatial query endpoints for checking flight permissions, retrieving restriction zones,
    and finding suitable launch sites.
    
    **Data Sources**: UK Civil Aviation Authority (CAA), NATS AIS, UK NOTAM Service  
    **License**: Data licensed under Open Government License v3.0  
    **Attribution**: Contains public sector information licensed under the OGL v3.0
  version: 1.0.0
  contact:
    name: DroneGo API Support
    email: api@dronego.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.dronego.example.com/v1
    description: Production server
  - url: https://staging-api.dronego.example.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
paths:
  /zones:
    get:
      summary: Query restriction zones by bounding box
      description: |
        Retrieve all active restriction zones within a specified geographic bounding box.
        Returns permanent restriction zones and active temporary restrictions (NOTAMs).
        Used for populating map viewport with restriction polygons.
        
        **Performance**: Typical response time <100ms for standard viewport.
        **Functional Requirement**: FR-002, FR-003, FR-004, FR-006
      tags:
        - zones
      parameters:
        - name: bounds
          in: query
          required: true
          description: Bounding box coordinates as minLon,minLat,maxLon,maxLat (WGS84)
          schema:
            type: string
            pattern: '^-?\d+\.?\d*,-?\d+\.?\d*,-?\d+\.?\d*,-?\d+\.?\d*$'
            example: "-0.5,51.4,0.0,51.6"
        - name: types
          in: query
          required: false
          description: Filter by zone types (comma-separated)
          schema:
            type: string
            example: "no-fly,controlled-airspace"
        - name: includeExpired
          in: query
          required: false
          description: Include expired temporary restrictions (default false)
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Successful response with GeoJSON FeatureCollection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestrictionZoneCollection'
              example:
                type: "FeatureCollection"
                metadata:
                  query_time_ms: 45
                  result_count: 12
                  bounds: [-0.5, 51.4, 0.0, 51.6]
                  data_freshness:
                    last_caa_sync: "2026-02-17T08:00:00Z"
                    last_notam_sync: "2026-02-17T10:30:00Z"
                features:
                  - type: "Feature"
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    geometry:
                      type: "Polygon"
                      coordinates: [[[-0.4543, 51.4700], [-0.4543, 51.4850], [-0.4200, 51.4850], [-0.4200, 51.4700], [-0.4543, 51.4700]]]
                    properties:
                      zone_type: "airport-frz"
                      restriction_name: "Heathrow FRZ"
                      authority_source: "CAA"
                      altitude_floor: 0
                      altitude_ceiling: 2500
                      description: "Flight Restriction Zone around Heathrow Airport. No unauthorized UAS operations permitted."
                      authorization_possible: false
                      confidence_level: "primary-authority"
                      restriction_category: "permanent"
                      last_updated: "2026-02-15T10:30:00Z"
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Invalid bounds format"
                message: "Bounds parameter must be in format: minLon,minLat,maxLon,maxLat"
                code: "INVALID_BOUNDS"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /zones/{zoneId}:
    get:
      summary: Get detailed information for a specific restriction zone
      description: |
        Retrieve complete details for a single restriction zone including full geometry,
        altitude restrictions, effective dates, and authorization information.
        
        **Functional Requirement**: FR-012 (detail panel)
      tags:
        - zones
      parameters:
        - name: zoneId
          in: path
          required: true
          description: UUID of the restriction zone
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response with zone details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestrictionZoneDetail'
        '404':
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin endpoints
  schemas:
    RestrictionZoneCollection:
      type: object
      properties:
        type:
          type: string
          enum: [FeatureCollection]
        metadata:
          type: object
          properties:
            query_time_ms:
              type: integer
              description: Query execution time in milliseconds
            result_count:
              type: integer
            bounds:
              type: array
              items:
                type: number
              minItems: 4
              maxItems: 4
            data_freshness:
              $ref: '#/components/schemas/DataFreshness'
        features:
          type: array
          items:
            $ref: '#/components/schemas/RestrictionZoneFeature'
    RestrictionZoneFeature:
      type: object
      properties:
        type:
          type: string
          enum: [Feature]
        id:
          type: string
          format: uuid
        geometry:
          $ref: '#/components/schemas/Polygon'
        properties:
          type: object
          properties:
            zone_type:
              type: string
              enum: [no-fly, controlled-airspace, military-zone, temporary-restriction, airport-frz, danger-area]
            restriction_name:
              type: string
            authority_source:
              type: string
            altitude_floor:
              type: integer
              nullable: true
            altitude_ceiling:
              type: integer
              nullable: true
            description:
              type: string
            authorization_possible:
              type: boolean
            confidence_level:
              type: string
              enum: [primary-authority, secondary-source, unverified]
            restriction_category:
              type: string
              enum: [permanent, temporary]
            effective_start:
              type: string
              format: date-time
              nullable: true
            effective_end:
              type: string
              format: date-time
              nullable: true
            last_updated:
              type: string
              format: date-time
    RestrictionZoneDetail:
      allOf:
        - $ref: '#/components/schemas/RestrictionZoneFeature'
        - type: object
          properties:
            properties:
              type: object
              properties:
                authorization_contact:
                  type: string
                  nullable: true
                data_source:
                  type: object
                  properties:
                    authority_name:
                      type: string
                    reliability_level:
                      type: string
                    license:
                      type: string
                    attribution:
                      type: string
...
    DataFreshness:
      type: object
      properties:
        last_caa_sync:
          type: string
          format: date-time
        last_nats_sync:
          type: string
          format: date-time
          nullable: true
        last_notam_sync:
          type: string
          format: date-time
        oldest_data_age_hours:
          type: number
        is_stale:
          type: boolean
          description: true if any data source is >48 hours old (FR-019)
    Polygon:
      type: object
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties: true
      required:
        - error
        - message
        - code